Fix the logic around remkl.
- There were the confusions about the remkl around the code:
sometimes it was used as a total klingons count, sometimes - as
a count of only an ordinary klingons. I added the separate
counters for those. Also removed the killed counters; added
the macros that calculates "killed" values as initial_val-remaining_val.
- Reset the klingons, starbases and supernovas on map upon an
initialization. This fixes the bug where the unkilled klingons
could propagate themselves to the subsequent game round.